refactor(hitobjects): change time-related properties from int to double
Update StartTime, Offset, and related properties to use double precision for better timing accuracy. Also clean up default value initializations and serialization logic.

+++ Coosu.Beatmap/Sections/HitObject/SliderInfo.cs
@@ -26,7 +26,7 @@ public class SliderInfo : SerializeWritableObject
     public Vector2 StartPoint { get; set; }
     public Vector2 EndPoint => ControlPoints[ControlPoints.Count - 1];
 
-    public int StartTime { get; set; }
+    public double StartTime { get; set; }
 
     public RawHitObject BaseObject { get; }
